requirements.txt
streamlit>=1.4.0
numpy>=1.23.0
pandas
matplotlib
langdetect
seaborn
plotly
boto3
spellchecker
bs4
fsspec
s3fs
opencv-python
keras>=2.6.0
tensorflow>=2.6.0
joblib>=1.1.0
nltk>=3.6.7
python-dotenv
protobuf==3.20.3
grpcio==1.48.2
wrapt>=1.15.0
altair==4.2.0
click==8.0.0